requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorful==0.5.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.67.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ray==2.39.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smart-open==7.0.5
sympy==1.13.1
tensorboardX==2.6.2.2
textual==0.85.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ultralytics==8.3.32
ultralytics-thop==2.0.11
urllib3==2.2.3
virtualenv==20.27.1
wrapt==1.16.0
yarl==1.17.1
